Add basic SVR solution. Improve CLI.
Add a basic SVR solution. Needs to be changed to accept a set of parameters from a previous learning session.

Add 'method' parameter to the CLI. For now options are 'rr1' (from the sample solution) and 'SVR'.

Note that both methods currently make an estimate, and use the sigmoid function to wrap a probability distribution around it.